fix for 64bit.
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@19379 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 63c001128487c8865b07828f0dd6056da76dfb5b

+++ stacks/visual_feature_detectors/outlet_detection/src/generalized_hough.cpp
@@ -225,7 +225,7 @@ CvMatND* buildHoughHistSparse(vector<feature_t>& input, const vector<feature_t>&
 					idx_[3] = i4;
 					sparse = cvCreateSparseMat(2,sizesxy,CV_32FC1);
 					CvScalar s;
-					s.val[0] = (int)sparse;
+					s.val[0] = (size_t)sparse;
 					cvSetND(mat,idx_,s);
 				}
 				int64 time1 = cvGetTickCount();
